test (pyutils): Separate out implementation and tests for miller_rabin functionality
